Further SDCARD DMA cleanup for other targets

+++ src/main/flight/mixer.c
@@ -102,7 +102,7 @@ void pgResetFn_motorConfig(motorConfig_t *motorConfig)
     motorConfig->dev.ioTags[3] = IO_TAG(MOTOR4_PIN);
 #else
     int motorIndex = 0;
-    for (int i = 0; i < USABLE_TIMER_CHANNEL_COUNT && motorIndex < MAX_SUPPORTED_MOTORS; i++) {
+    for (int i = 0; i < (int)USABLE_TIMER_CHANNEL_COUNT && motorIndex < MAX_SUPPORTED_MOTORS; i++) {
         if (timerHardware[i].usageFlags & TIM_USE_MOTOR) {
             motorConfig->dev.ioTags[motorIndex] = timerHardware[i].tag;
             motorIndex++;
